Multi-page vectored IOs
New fox option: nppas

'nppas' defines the maximum number of ppas per IO command, Fox will
create multi-page commands when possible. It depends of the engine and
how the IOs are scheduled. For now, multi-pages are set when the IO
sequence is in the same block and same LUN.

Engines:
1 - fox-sequential: Up to 64 ppas per command;
2 - fox-round-robin: Only 1 page(planes*sectors_per_page ppas) per command;