Fox 1.2 - Multi-type buffer and full corruption detection
- Three buffer types (-m option):
 -m 1 -> random data
 -m 2 -> human readable data
 -m 3 -> geometry based data (required for engine 3 and 100% reads)

All types perform corruption check
All engines and cases will check for corruption

If -m is not present, the data will be type random data and fox
will NOT check for corruption.

limitation: Engine 2 with human readable data: Only the page is
updated in the readable buffer. Channel, LUN and block will be 0
for all blocks in the device. Other engines will works with full
geometry.